--- a/pkg/ttl/client/BUILD.bazel
+++ b/pkg/ttl/client/BUILD.bazel
@@ -10,6 +10,7 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//pkg/ddl/util",
+        "//pkg/util/intest",
         "//pkg/util/logutil",
         "@com_github_google_uuid//:uuid",
         "@com_github_pingcap_errors//:errors",