--- a/pkg/ttl/ttlworker/job_manager_test.go
+++ b/pkg/ttl/ttlworker/job_manager_test.go
@@ -318,7 +318,7 @@ func TestOnTimerTick(t *testing.T) {
 	require.Equal(t, now, syncTime)
 
 	// resume after a very short duration
-	now = now.Add(time.Second)
+	now = now.Add(time.Microsecond * 999)
 	se.sessionInfoSchema = newMockInfoSchemaWithVer(101, tbl.TableInfo)
 	m.onTimerTick(se, rt, syncer, now)
 	require.Same(t, innerRT, rt.rt)
@@ -326,10 +326,10 @@ func TestOnTimerTick(t *testing.T) {
 	require.Equal(t, 1, len(syncer.key2Timers))
 	syncTime, syncVer = syncer.GetLastSyncInfo()
 	require.Equal(t, int64(100), syncVer)
-	require.Equal(t, now.Add(-time.Second), syncTime)
+	require.Equal(t, now.Add(-999*time.Microsecond), syncTime)
 
 	// resume after a middle duration
-	now = now.Add(6 * time.Second)
+	now = now.Add(2 * time.Millisecond)
 	m.onTimerTick(se, rt, syncer, now)
 	require.Same(t, innerRT, rt.rt)
 	require.True(t, innerRT.Running())