<!--

Thank you for contributing to TiDB!

PR Title Format:
1. pkg [, pkg2, pkg3]: what's changed
2. *: what's changed

-->

### What problem does this PR solve?
<!--

Please create an issue first to describe the problem.

There MUST be one line starting with "Issue Number:  " and
linking the relevant issues via the "close" or "ref".

For more info, check https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/contribute-code.html#referring-to-an-issue.

-->

Issue Number: close #56718

### What changed and how does it work?

1. Provides `boostJobScheduleForTest` to make some intervals smaller to boost test.
2. Modify some cases to make it more reasonable and faster

In master, run tests in `pkg/ttl/ttlworker` takes 34.5s:

```
> X="run pkg/ttl/ttlworker" make ut
cd tools/check; \
	GO111MODULE=on go build -o ../bin/ut ut.go
cd tools/check; \
	GO111MODULE=on go build -o ../bin/xprog xprog.go
GOBIN=/Users/wangchao/Code/pingcap/tidb/tools/bin GO111MODULE=on go install github.com/pingcap/failpoint/failpoint-ctl@9b3b6e3
tools/bin/ut run pkg/ttl/ttlworker || { find $PWD/ -type d | grep -vE "(\.git|tools)" | xargs tools/bin/failpoint-ctl disable; exit 1; }
2024/10/18 12:09:12 maxprocs: Leaving GOMAXPROCS=10: CPU quota undefined
building task finish, parallelism=20, count=55, takes=1m3.719834292s
run all tasks takes 34.5330225s
```

In this PR, it takes 3.4s:

```
> X="run pkg/ttl/ttlworker" make ut
cd tools/check; \
	GO111MODULE=on go build -o ../bin/ut ut.go
cd tools/check; \
	GO111MODULE=on go build -o ../bin/xprog xprog.go
GOBIN=/Users/wangchao/Code/pingcap/tidb/tools/bin GO111MODULE=on go install github.com/pingcap/failpoint/failpoint-ctl@9b3b6e3
tools/bin/ut run pkg/ttl/ttlworker || { find $PWD/ -type d | grep -vE "(\.git|tools)" | xargs tools/bin/failpoint-ctl disable; exit 1; }
2024/10/18 12:11:16 maxprocs: Leaving GOMAXPROCS=10: CPU quota undefined
building task finish, parallelism=20, count=55, takes=30.296769208s
run all tasks takes 3.442459375s
```

### Check List

Tests <!-- At least one of them must be included. -->

- [x] Unit test
- [ ] Integration test
- [ ] Manual test (add detailed scripts or steps below)
- [ ] No need to test
  > - [ ] I checked and no code files have been changed.
  > <!-- Or your custom  "No need to test" reasons -->

Side effects

- [ ] Performance regression: Consumes more CPU
- [ ] Performance regression: Consumes more Memory
- [ ] Breaking backward compatibility

Documentation

- [ ] Affects user behaviors
- [ ] Contains syntax changes
- [ ] Contains variable changes
- [ ] Contains experimental features
- [ ] Changes MySQL compatibility

### Release note

<!-- compatibility change, improvement, bugfix, and new feature need a release note -->

Please refer to [Release Notes Language Style Guide](https://pingcap.github.io/tidb-dev-guide/contribute-to-tidb/release-notes-style-guide.html) to write a quality release note.

```release-note
None
```